Widen the session id to 64 bits
NEXT_SESSION_ID was an AtomicI32 incremented once per pipe connection
and never wrapped deliberately, so after ~2^31 connections it would go
negative and eventually collide with a live session, cross-wiring two
applications' composing state. Reaching that takes billions of
connections in one server lifetime, so this is about closing the hole,
not about a bug anyone hits.

The id crosses the FFI, so widening it is an ABI change in three places
that nothing checks mechanically: ffi.h (the canonical declaration), the
Rust extern block mirroring it, and the Swift @_cdecl exports plus the
`sessions` dictionary key. All three now say 64-bit, along with the
Rust-side counter, PipeConnectInfo, and the idle-eviction map.

Building proves nothing here. On the x64 calling convention a 64-bit
argument read as 32 bits still yields the right value for small ids, so
a half-applied widening would look perfectly healthy and only fail past
2^32 — exactly the range this change exists to protect, and one no
ordinary test reaches. Verified instead by temporarily stepping the
counter by 2^32, so consecutive connections got ids identical in their
low 32 bits: under truncation both sessions would collapse into one and
the interleaved typing in sessions_are_isolated_between_connections
would mix. All eight live-engine smoke tests passed with those ids, and
again after restoring the normal counter.

64 bits is not literally unbounded, but at one connection per
millisecond it is some 300 million years.

Closes https://github.com/almandite1/azooKey-Windows/issues/2

+++ crates/server/src/ffi.rs
@@ -13,7 +13,9 @@
 //!   in main())
 //! - out-parameters are 32-bit (c_int on both sides)
 //! - `session` selects the per-client composing state; it is the pipe
-//!   connection id assigned in lib.rs (see PipeConnectInfo)
+//!   connection id assigned in lib.rs (see PipeConnectInfo). 64-bit, so the
+//!   monotonic counter behind it cannot realistically wrap onto a live
+//!   session
 //! - every returned string/list is owned by the callee and must be handed
 //!   back to FreeString / FreeComposedText after copying
 
@@ -29,18 +31,18 @@ pub(crate) struct FFICandidate {
 
 unsafe extern "C" {
     pub(crate) fn Initialize(path: *const c_char);
-    pub(crate) fn SetContext(session: c_int, context: *const c_char);
+    pub(crate) fn SetContext(session: i64, context: *const c_char);
     pub(crate) fn AppendText(
-        session: c_int,
+        session: i64,
         input: *const c_char,
         cursorPtr: *mut c_int,
     ) -> *mut c_char;
-    pub(crate) fn RemoveText(session: c_int, cursorPtr: *mut c_int) -> *mut c_char;
-    pub(crate) fn MoveCursor(session: c_int, offset: c_int, cursorPtr: *mut c_int) -> *mut c_char;
-    pub(crate) fn ShrinkText(session: c_int, offset: c_int) -> *mut c_char;
-    pub(crate) fn ClearText(session: c_int);
-    pub(crate) fn GetComposedText(session: c_int, lengthPtr: *mut c_int) -> *mut *mut FFICandidate;
-    pub(crate) fn RemoveSession(session: c_int);
+    pub(crate) fn RemoveText(session: i64, cursorPtr: *mut c_int) -> *mut c_char;
+    pub(crate) fn MoveCursor(session: i64, offset: c_int, cursorPtr: *mut c_int) -> *mut c_char;
+    pub(crate) fn ShrinkText(session: i64, offset: c_int) -> *mut c_char;
+    pub(crate) fn ClearText(session: i64);
+    pub(crate) fn GetComposedText(session: i64, lengthPtr: *mut c_int) -> *mut *mut FFICandidate;
+    pub(crate) fn RemoveSession(session: i64);
     pub(crate) fn LoadConfig();
     pub(crate) fn FreeString(ptr: *mut c_char);
     pub(crate) fn FreeComposedText(listPtr: *mut *mut FFICandidate, length: c_int);
